Removed interface RejReasonDAO
